[qca-nss-dp] Enable low threshold interrupts for Rx-fill rings
In current DP design, the Rx fill ring is replenished as part
of Rx Desc ring process. In some cases when system is running
low on memory, it is feasible that the buffer allocation for
Rx-Fill ring continuously fail and empty buffers in rx-fill ring
keep depleting. Once the number of rx-fill buffers goes below a
certain threshold, EDMA asserts the flow control to the PPE switch
core. If the EDMA Rx descriptors rings are empty, then there will
not be any interrupts generated unless new empty buffers are produced
which is done only in rx-reap. So this results in a deadlock between
fill and reap. To overcome this deadlock, enable low threshold
interrupt for Rx-fill ring to replenish minimum buffers in rx-fill
to keep system operational.

Change-Id: Id8ab320b8d0c312c80ef614b49b4046db6fe7777
Signed-off-by: Amit Gupta <quic_amitgupt@quicinc.com>