Revert^2 "[qca-nss-dp] Enable EDMA driver to allocate from beyond 4GB space."
This reverts commit 5c6b38d356b0c1fb481b34fd8a5cc864576938cb.

Reason for revert:

Initially the gerrit was first merged on component branch, NSSDevtest observed some performance impact and we had to revert the gerrit.
Issue was fixed and we then validated and merged the change on nss branch

We also need to get the same changes on win.nss.1.0 branch. And hence we need to undo the revert.

1)	win.nss.1.0 branch: [qca-nss-dp] Enable EDMA driver to allocate from beyond 4GB space. (5421761) · Gerrit Code Review (quicinc.com)
2)	revert of win.nss.1.0 branch: Revert "[qca-nss-dp] Enable EDMA driver to allocate from beyond 4GB space." (5429015) · Gerrit Code Review (quicinc.com)
3)	undo the revert on 2nd step. : Revert^2 "[qca-nss-dp] Enable EDMA driver to allocate from beyond 4GB space." (5519924) · Gerrit Code Review (quicinc.com)

Change-Id: Iba487eced56d04a61781f0c491ba809ad7fa65d1
Signed-off-by: Aniruddha Bhat Anemajalu <quic_aanemaja@quicinc.com>