[qca-nss-dp] EDMA driver base re-architecture.
1. Add infrastructure to read the per type (SFE/VP modes, etc) ring/queue
mapping, setup information from module params instead of DTSI and to
store into per type data structures to be used for ring allocation
and setup later.

2. Allocate and setup the TX/RX rings, queues for different modes as per
the information parsed from module params making it flexible to add, remove
rings as required.

3. Control path changes catering to the above implementation regarding
queue setup, debugfs changes, setting up ring resources(ex. NAPI, IRQ, etc)
based on the rings that are to be used.

4. Store the IRQ numbers for all the rings at once and then request irqs
based on the rings that are enabled. This makes it easy to add a new ring
since the IRQ numbers are already parsed from DTSI.

Change-Id: Iecd4bcc1dcff41a76c2bfe5d8283304cc7e5ee3d
Signed-off-by: Parikshit Gune <pgune@qti.qualcomm.com>